Use tablebase only for comment evaluation in Practice with computer
The "best" move in the tablebase results is often a move that is far from
the best move to make.
For example the top move can be a mate in 2 instead of a mate in 20.
This makes the Practice with computer feature much less useful than it should be.

This commit changes the behavior to use the tablebase only for the accuracy
comment.